+++ src/components/form-tabs/types.ts
@@ -1,0 +1,7 @@
+// eslint-disable-next-line import/prefer-default-export
+export enum TabState {
+  valid = 'valid',
+  error = 'error',
+  default = 'default',
+  optional = 'optional',
+}
