+++ src/components/row/button/index.tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react';
 // Components
 import { Button } from 'rebass';
-// eslint-disable-next-line import/no-cycle
 import Tooltip from '../../tooltip';
 import rowButtonStyles from './row-button.styles';
